rename - #5: User, Exchange 도메인 패키지 경로 수정, getExchanges 메서드 RequestParam 기본값 설정
